+++ src/scope.rs
@@ -175,7 +175,7 @@ impl<'a> Scope<'a> {
     ///
     /// Modules are used for accessing member variables, functions and plugins under a namespace.
     #[cfg(not(feature = "no_module"))]
-    pub fn push_module<K: Into<Cow<'a, str>>>(&mut self, name: K, mut value: Module) {
+    pub fn push_module<K: Into<Cow<'a, str>>>(&mut self, name: K, value: Module) {
         self.push_module_internal(name, value);
     }